requirements.txt
pillow
piexif
tk
tensorflow
numpy
opencv-python
matplotlib
scikit-learn
joblib